+++ src/main/java/reisetech/StudentManagement/StudentManagementApplication.java
@@ -6,26 +6,16 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import reisetech.StudentManagement.StudentRepository.StudentRepository;
+import reisetech.StudentManagement.data.Student;
+import reisetech.StudentManagement.data.StudentCourse;
 
 @SpringBootApplication
-@RestController
-public class StudentManagementApplication {
 
-  @Autowired
-  private StudentRepository repository;
+public class StudentManagementApplication {
 
   public static void main(String[] args) {
     SpringApplication.run(StudentManagementApplication.class, args);
-  }
-
-  @GetMapping("/studentList")
-  public List<Student> getStudentList() {
-    return repository.searchStudent();
-  }
-
-  @GetMapping("/studentCourseList")
-  public List<StudentCourse> getStudentCourseList() {
-    return repository.searchStudentCourse();
   }
 }
