+++ src/main/java/reisetech/StudentManagement/controller/StudentController.java
@@ -1,0 +1,31 @@
+package reisetech.StudentManagement.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import reisetech.StudentManagement.data.Student;
+import reisetech.StudentManagement.data.StudentCourse;
+import reisetech.StudentManagement.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  private StudentService service;
+
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+  }
+
+  @GetMapping("/studentCourseList")
+  public List<StudentCourse> getStudentCourseList() {
+    return service.searchStudentCourseList();
+  }
+}
